bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = "Telegram Bot for Habit Tracker"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^22.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"
apscheduler = "^3.8.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"